+++ core/src/main/java/org/infinispan/configuration/cache/FileCacheStoreConfiguration.java
@@ -4,6 +4,7 @@ import org.infinispan.commons.configuration.BuiltBy;
 import org.infinispan.commons.configuration.ConfigurationFor;
 import org.infinispan.commons.util.TypedProperties;
 import org.infinispan.configuration.cache.FileCacheStoreConfigurationBuilder.FsyncMode;
+import org.infinispan.loaders.file.DelegateFileCacheStore;
 import org.infinispan.loaders.file.FileCacheStore;
 
 /**
@@ -13,16 +14,18 @@ import org.infinispan.loaders.file.FileCacheStore;
  * @since 5.1
  */
 @BuiltBy(FileCacheStoreConfigurationBuilder.class)
-@ConfigurationFor(FileCacheStore.class)
+@ConfigurationFor(DelegateFileCacheStore.class)
 public class FileCacheStoreConfiguration extends AbstractLockSupportStoreConfiguration {
 
    private final String location;
    private final long fsyncInterval;
    private final FsyncMode fsyncMode;
    private final int streamBufferSize;
+   private final int maxEntries;
+   private final boolean deprecatedBucketFormat;
 
-   FileCacheStoreConfiguration(String location, long fsyncInterval,
-         FsyncMode fsyncMode, int streamBufferSize, long lockAcquistionTimeout,
+   FileCacheStoreConfiguration(String location, int maxEntries, boolean deprecatedBucketFormat,
+         long fsyncInterval, FsyncMode fsyncMode, int streamBufferSize, long lockAcquistionTimeout,
          int lockConcurrencyLevel, boolean purgeOnStartup, boolean purgeSynchronously,
          int purgerThreads, boolean fetchPersistentState, boolean ignoreModifications,
          TypedProperties properties, AsyncStoreConfiguration async,
@@ -31,15 +34,19 @@ public class FileCacheStoreConfiguration extends AbstractLockSupportStoreConfigu
             purgeSynchronously, purgerThreads, fetchPersistentState,
             ignoreModifications, properties, async, singletonStore);
       this.location = location;
+      this.maxEntries = maxEntries;
+      this.deprecatedBucketFormat = deprecatedBucketFormat;
       this.fsyncInterval = fsyncInterval;
       this.fsyncMode = fsyncMode;
       this.streamBufferSize = streamBufferSize;
    }
 
+   @Deprecated
    public long fsyncInterval() {
       return fsyncInterval;
    }
 
+   @Deprecated
    public FsyncMode fsyncMode() {
       return fsyncMode;
    }
@@ -48,6 +55,15 @@ public class FileCacheStoreConfiguration extends AbstractLockSupportStoreConfigu
       return location;
    }
 
+   public boolean deprecatedBucketFormat() {
+      return deprecatedBucketFormat;
+   }
+
+   public int maxEntries() {
+      return maxEntries;
+   }
+
+   @Deprecated
    public int streamBufferSize() {
       return streamBufferSize;
    }
@@ -57,6 +73,8 @@ public class FileCacheStoreConfiguration extends AbstractLockSupportStoreConfigu
       return "FileCacheStoreConfiguration{" +
             "fsyncInterval=" + fsyncInterval +
             ", location='" + location + '\'' +
+            ", maxEntries=" + maxEntries +
+            ", deprecatedBucketFormat=" + deprecatedBucketFormat +
             ", fsyncMode=" + fsyncMode +
             ", streamBufferSize=" + streamBufferSize +
             ", lockAcquistionTimeout=" + lockAcquistionTimeout() +
@@ -85,6 +103,8 @@ public class FileCacheStoreConfiguration extends AbstractLockSupportStoreConfigu
       if (fsyncMode != that.fsyncMode) return false;
       if (location != null ? !location.equals(that.location) : that.location != null)
          return false;
+      if (maxEntries != that.maxEntries) return false;
+      if (deprecatedBucketFormat != that.deprecatedBucketFormat) return false;
 
       return true;
    }
@@ -93,9 +113,12 @@ public class FileCacheStoreConfiguration extends AbstractLockSupportStoreConfigu
    public int hashCode() {
       int result = super.hashCode();
       result = 31 * result + (location != null ? location.hashCode() : 0);
+      result = 31 * result + maxEntries;
+      result = 31 * result + (deprecatedBucketFormat ? 1 : 0);
       result = 31 * result + (int) (fsyncInterval ^ (fsyncInterval >>> 32));
       result = 31 * result + (fsyncMode != null ? fsyncMode.hashCode() : 0);
       result = 31 * result + streamBufferSize;
       return result;
    }
+
 }
