+++ core/src/main/java/org/infinispan/configuration/parsing/Parser60.java
@@ -559,9 +559,6 @@ public class Parser60 implements ConfigurationParser {
             case FILE_STORE:
                parseFileStore(reader, holder);
                break;
-            case SINGLE_FILE_STORE:
-               parseSingleFileStore(reader, holder);
-               break;
             case LOADER:
                parseLoader(reader, holder);
                break;
@@ -572,28 +569,6 @@ public class Parser60 implements ConfigurationParser {
                reader.handleAny(holder);
          }
       }
-   }
-
-   private void parseSingleFileStore(XMLExtendedStreamReader reader, ConfigurationBuilderHolder holder) throws XMLStreamException {
-      ConfigurationBuilder builder = holder.getCurrentConfigurationBuilder();
-      SingleFileCacheStoreConfigurationBuilder storeBuilder = builder.loaders().addSingleFileCacheStore();
-      for (int i = 0; i < reader.getAttributeCount(); i++) {
-         ParseUtils.requireNoNamespaceAttribute(reader, i);
-         String value = replaceProperties(reader.getAttributeValue(i));
-         Attribute attribute = Attribute.forName(reader.getAttributeLocalName(i));
-         switch (attribute) {
-            case LOCATION:
-               storeBuilder.location(value);
-               break;
-            case MAX_ENTRIES:
-               storeBuilder.maxEntries(Integer.parseInt(value));
-               break;
-            default:
-               parseCommonLoaderAttributes(reader, i, storeBuilder);
-               break;
-         }
-      }
-      parseStoreChildren(reader, storeBuilder);
    }
 
    private void parseClusterLoader(XMLExtendedStreamReader reader, ConfigurationBuilderHolder holder) throws XMLStreamException {
@@ -625,6 +600,12 @@ public class Parser60 implements ConfigurationParser {
          switch (attribute) {
          case LOCATION:
             fcscb.location(value);
+            break;
+         case MAX_ENTRIES:
+            fcscb.maxEntries(Integer.parseInt(value));
+            break;
+         case DEPRECATED_BUCKET_FORMAT:
+            fcscb.deprecatedBucketFormat(Boolean.parseBoolean(value));
             break;
          case FSYNC_INTERVAL:
             fcscb.fsyncInterval(Long.parseLong(value));
