+++ core/src/test/java/org/infinispan/loaders/file/BucketFileCacheStorePeriodicTest.java
@@ -4,7 +4,7 @@ import org.infinispan.configuration.cache.FileCacheStoreConfigurationBuilder;
 import org.testng.annotations.Test;
 
 @Test(groups = "unit", testName = "loaders.file.FileCacheStorePeriodicTest")
-public class FileCacheStorePeriodicTest extends FileCacheStoreTest {
+public class BucketFileCacheStorePeriodicTest extends BucketFileCacheStoreTest {
 
    @Override
    protected FileCacheStoreConfigurationBuilder.FsyncMode getFsyncMode() {
